+++ util/opentelemetry-util-genai/src/opentelemetry/util/genai/_upload/completion_hook.py
@@ -15,11 +15,9 @@
 
 from __future__ import annotations
 
-import json
 import logging
 import posixpath
 import threading
-from base64 import b64encode
 from concurrent.futures import Future, ThreadPoolExecutor
 from contextlib import ExitStack
 from dataclasses import asdict, dataclass
@@ -39,6 +37,7 @@ from opentelemetry.util.genai.completion_hook import CompletionHook
 from opentelemetry.util.genai.environment_variables import (
     OTEL_INSTRUMENTATION_GENAI_UPLOAD_FORMAT,
 )
+from opentelemetry.util.genai.utils import gen_ai_json_dump
 
 GEN_AI_INPUT_MESSAGES_REF: Final = (
     gen_ai_attributes.GEN_AI_INPUT_MESSAGES + "_ref"
@@ -60,9 +59,9 @@ _logger = logging.getLogger(__name__)
 
 @dataclass
 class Completion:
-    inputs: list[types.InputMessage]
-    outputs: list[types.OutputMessage]
-    system_instruction: list[types.MessagePart]
+    inputs: list[types.InputMessage] | None
+    outputs: list[types.OutputMessage] | None
+    system_instruction: list[types.MessagePart] | None
 
 
 @dataclass
@@ -192,12 +191,7 @@ class UploadCompletionHook(CompletionHook):
 
         with self._fs.open(path, "w", content_type=content_type) as file:
             for message in message_lines:
-                json.dump(
-                    message,
-                    file,
-                    separators=(",", ":"),
-                    cls=Base64JsonEncoder,
-                )
+                gen_ai_json_dump(message, file)
                 file.write("\n")
 
     def on_completion(
@@ -210,10 +204,13 @@ class UploadCompletionHook(CompletionHook):
         log_record: LogRecord | None = None,
         **kwargs: Any,
     ) -> None:
+        if not any([inputs, outputs, system_instruction]):
+            return
+        # An empty list will not be uploaded.
         completion = Completion(
-            inputs=inputs,
-            outputs=outputs,
-            system_instruction=system_instruction,
+            inputs=inputs or None,
+            outputs=outputs or None,
+            system_instruction=system_instruction or None,
         )
         # generate the paths to upload to
         ref_names = self._calculate_ref_path()
@@ -225,23 +222,36 @@ class UploadCompletionHook(CompletionHook):
         ) -> JsonEncodeable:
             return [asdict(dc) for dc in dataclass_list]
 
+        references = [
+            (ref_name, ref, ref_attr)
+            for ref_name, ref, ref_attr in [
+                (
+                    ref_names.inputs_ref,
+                    completion.inputs,
+                    GEN_AI_INPUT_MESSAGES_REF,
+                ),
+                (
+                    ref_names.outputs_ref,
+                    completion.outputs,
+                    GEN_AI_OUTPUT_MESSAGES_REF,
+                ),
+                (
+                    ref_names.system_instruction_ref,
+                    completion.system_instruction,
+                    GEN_AI_SYSTEM_INSTRUCTIONS_REF,
+                ),
+            ]
+            if ref
+        ]
         self._submit_all(
             {
-                # Use partial to defer as much as possible to the background threads
-                ref_names.inputs_ref: partial(to_dict, completion.inputs),
-                ref_names.outputs_ref: partial(to_dict, completion.outputs),
-                ref_names.system_instruction_ref: partial(
-                    to_dict, completion.system_instruction
-                ),
-            },
+                ref_name: partial(to_dict, ref)
+                for ref_name, ref, _ in references
+            }
         )
 
         # stamp the refs on telemetry
-        references = {
-            GEN_AI_INPUT_MESSAGES_REF: ref_names.inputs_ref,
-            GEN_AI_OUTPUT_MESSAGES_REF: ref_names.outputs_ref,
-            GEN_AI_SYSTEM_INSTRUCTIONS_REF: ref_names.system_instruction_ref,
-        }
+        references = {ref_attr: name for name, _, ref_attr in references}
         if span:
             span.set_attributes(references)
         if log_record:
@@ -265,10 +275,3 @@ class UploadCompletionHook(CompletionHook):
 
             # Queue is flushed and blocked, start shutdown
             self._executor.shutdown(wait=False)
-
-
-class Base64JsonEncoder(json.JSONEncoder):
-    def default(self, o: Any) -> Any:
-        if isinstance(o, bytes):
-            return b64encode(o).decode()
-        return super().default(o)