+++ util/opentelemetry-util-genai/src/opentelemetry/util/genai/utils.py
@@ -12,8 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import json
 import logging
 import os
+from base64 import b64encode
+from functools import partial
+from typing import Any
 
 from opentelemetry.instrumentation._semconv import (
     _OpenTelemetrySemanticConventionStability,
@@ -54,3 +58,23 @@ def get_content_capturing_mode() -> ContentCapturingMode:
             ", ".join(e.name for e in ContentCapturingMode),
         )
         return ContentCapturingMode.NO_CONTENT
+
+
+class _GenAiJsonEncoder(json.JSONEncoder):
+    def default(self, o: Any) -> Any:
+        if isinstance(o, bytes):
+            return b64encode(o).decode()
+        return super().default(o)
+
+
+gen_ai_json_dump = partial(
+    json.dump, separators=(",", ":"), cls=_GenAiJsonEncoder
+)
+"""Should be used by GenAI instrumentations when serializing objects that may contain
+bytes, datetimes, etc. for GenAI observability."""
+
+gen_ai_json_dumps = partial(
+    json.dumps, separators=(",", ":"), cls=_GenAiJsonEncoder
+)
+"""Should be used by GenAI instrumentations when serializing objects that may contain
+bytes, datetimes, etc. for GenAI observability."""