+++ util/opentelemetry-util-genai/CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3798](#3798))
 - Record content-type and use canonical paths in fsspec genai uploader
   ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3795](#3795))
+- Make inputs / outputs / system instructions optional params to `on_completion`,
+  ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3802](#3802)).
 
 ## Version 0.1b0 (2025-09-25)